Split resolvers and schemas
