+++ package.json
@@ -43,31 +43,34 @@
     "trailingComma": "all"
   },
   "dependencies": {
-    "apollo-server-koa": "^2.12.0",
-    "graphql": "^15.0.0",
-    "koa": "^2.11.0",
-    "koa-bodyparser": "^4.3.0",
-    "koa-json": "^2.0.2",
-    "koa-logger": "^3.2.1"
+    "apollo-server-koa": "~2.12.0",
+    "graphql": "~15.0.0",
+    "graphql-tools": "~5.0.0",
+    "koa": "~2.11.0",
+    "koa-bodyparser": "~4.3.0",
+    "koa-json": "~2.0.2",
+    "koa-logger": "~3.2.1",
+    "lodash.merge": "~4.6.2"
   },
   "devDependencies": {
-    "@types/jest": "^25.2.1",
-    "@types/koa": "^2.11.3",
-    "@types/koa-bodyparser": "^4.3.0",
-    "@types/koa-json": "^2.0.18",
-    "@types/koa-logger": "^3.1.1",
-    "@typescript-eslint/eslint-plugin": "^2.28.0",
-    "@typescript-eslint/parser": "^2.28.0",
-    "eslint": "^6.8.0",
-    "eslint-config-cheminfo": "^3.0.0",
-    "eslint-config-cheminfo-typescript": "^5.0.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jest": "^23.8.2",
-    "eslint-plugin-prettier": "^3.1.3",
-    "jest": "^25.3.0",
-    "prettier": "^2.0.4",
-    "rimraf": "^3.0.2",
-    "ts-jest": "^25.3.1",
-    "typescript": "^3.8.3"
+    "@types/jest": "~25.2.1",
+    "@types/koa": "~2.11.3",
+    "@types/koa-bodyparser": "~4.3.0",
+    "@types/koa-json": "~2.0.18",
+    "@types/koa-logger": "~3.1.1",
+    "@types/lodash.merge": "^4.6.6",
+    "@typescript-eslint/eslint-plugin": "~2.28.0",
+    "@typescript-eslint/parser": "~2.28.0",
+    "eslint": "~6.8.0",
+    "eslint-config-cheminfo": "~3.0.0",
+    "eslint-config-cheminfo-typescript": "~5.0.0",
+    "eslint-plugin-import": "~2.20.2",
+    "eslint-plugin-jest": "~23.8.2",
+    "eslint-plugin-prettier": "~3.1.3",
+    "jest": "~25.3.0",
+    "prettier": "~2.0.4",
+    "rimraf": "~3.0.2",
+    "ts-jest": "~25.3.1",
+    "typescript": "~3.8.3"
   }
 }
