--- src/__tests__/test.ts
@@ -1,7 +1,0 @@
-import { myModule } from '..';
-
-describe('test myModule', () => {
-  it('should return 42', () => {
-    expect(myModule()).toBe(42);
-  });
-});
