+++ src/schemas/author.ts
@@ -1,0 +1,17 @@
+import { gql } from 'apollo-server-koa';
+
+export const authorSchema = gql`
+  type Author {
+    id: Int!
+    firstName: String
+    lastName: String
+    """
+    the list of Posts by this author
+    """
+    posts: [Post]
+  }
+
+  extend type Query {
+    author(id: Int!): Author
+  }
+`;
